complete delete article functionality
