+++ app/components/ArticleDeleteButton.js
@@ -1,0 +1,11 @@
+var React = require('react');
+
+var ArticleDeleteButton = React.createClass({
+  render: function () {
+    return (
+      <button type="button" onClick={ this.props.handleDeleteArticle }>delete</button>
+    )
+  }
+});
+
+module.exports = ArticleDeleteButton;
