+++ app/components/TeamShowHeader.js
@@ -2,10 +2,9 @@ var React = require('react');
 
 var TeamShowHeader = React.createClass({
   render: function () {
-	// debugger;
     return (
       <div>
-        {/*<h1>{ this.props.team.name }</h1>*/}
+        {<h1>{ this.props.team.name }</h1>}
       </div>
     )
   }
