+++ app/containers/ArticleEditContainer.js
@@ -4,14 +4,18 @@ var $ = require('jquery');
 
 var ArticleEditContainer = React.createClass({
   renderButton: function () {
-    return <button id="edit-article-button" type="button" onClick={ this.props.handleToggleEditArticleForm }>Edit</button>
+    return (
+      <button id="edit-article-button" type="button" onClick={ this.props.handleToggleEditArticleForm }>Edit</button>
+    )
   },
   renderForm: function () {
-    return <EditArticleFormContainer
+    return (
+      <EditArticleFormContainer
         articleBody={ this.props.articleBody }
         onSubmitArticleEdit={ this.props.onSubmitArticleEdit }
         onUpdateArticle={ this.props.handleUpdateArticle }
         onToggleEditArticleForm={ this.props.handleToggleEditArticleForm } />
+      )
   },
   render: function () {
     if ( this.props.editing ) {
